get channel creation and channel page fetching to work

+++ server/Services/Services/Channels/ChannelMemberService.cs
@@ -36,13 +36,13 @@ public class ChannelMemberService(IConfiguration config) : DbService(config)
         return result;
     }
 
-    public async Task<int> GetMemberCount(Guid channelId)
+    public async Task<long> GetMemberCount(Guid channelId)
     {
-        int? result = (int?)await ExecuteScalarAsync($@"
+        long? result = (long?)await ExecuteScalarAsync($@"
             SELECT COUNT(*) FROM channelmembers
             WHERE channelId = '{channelId}'::uuid
         ");
-        return result ?? 0;
+        return result ?? 0; 
     }
 
     public async Task Delete(ChannelMember channelMember)
